requirements.txt
flake8
PyYAML
colorama